Alignment for v1 and v2 interfaces for parameters and SOLID principles
