+++ MSAL/src/native_auth/controllers/factories/MSALNativeAuthControllerFactory.swift
@@ -28,7 +28,7 @@ protocol MSALNativeAuthControllerBuildable {
     func makeJITController(cacheAccessor: MSALNativeAuthCacheInterface) -> MSALNativeAuthJITControlling
     func makeResetPasswordController(cacheAccessor: MSALNativeAuthCacheInterface) -> MSALNativeAuthResetPasswordControlling
     func makeCredentialsController(cacheAccessor: MSALNativeAuthCacheInterface) -> MSALNativeAuthCredentialsControlling
-    func makeV2FlowController(cacheAccessor: MSALNativeAuthCacheInterface) -> MSALNativeAuthV2FlowControlling
+    func makeV2FlowController(cacheAccessor: MSALNativeAuthCacheInterface) -> MSALNativeAuthFlowControlling
 }
 
 final class MSALNativeAuthControllerFactory: MSALNativeAuthControllerBuildable {
@@ -58,7 +58,7 @@ final class MSALNativeAuthControllerFactory: MSALNativeAuthControllerBuildable {
         return MSALNativeAuthCredentialsController(config: config, cacheAccessor: cacheAccessor)
     }
 
-    func makeV2FlowController(cacheAccessor: MSALNativeAuthCacheInterface) -> MSALNativeAuthV2FlowControlling {
-        return MSALNativeAuthV2FlowController(config: config, cacheAccessor: cacheAccessor)
+    func makeV2FlowController(cacheAccessor: MSALNativeAuthCacheInterface) -> MSALNativeAuthFlowControlling {
+        return MSALNativeAuthFlowController(config: config, cacheAccessor: cacheAccessor)
     }
 }
