+++ MSAL/src/native_auth/network/v2/MSALNativeAuthV2RequestConfigurator.swift
@@ -26,147 +26,6 @@ import Foundation
 
 @_implementationOnly import MSAL_Private
 
-// MARK: - Request parameters
-
-/// Describes a single V2 native-auth request: each concrete parameter type knows its target URL,
-/// HTTP method, body, body encoding and telemetry identity. `MSALNativeAuthV2RequestConfigurator`
-/// turns any of these into a fully-configured `MSIDHttpRequest` that reuses the shared AAD request
-/// pipeline (device-id headers, PkeyAuth, correlation, server telemetry).
-protocol MSALNativeAuthV2Requestable {
-    var context: MSALNativeAuthRequestContext { get }
-    var httpMethod: String { get }
-    var encoding: MSALNativeAuthUrlRequestEncoding { get }
-    var apiId: MSALNativeAuthTelemetryApiId { get }
-    var operationType: MSALNativeAuthOperationType { get }
-    /// `true` only for the `/token` endpoint, which returns a plain OAuth response (not HAL) and must
-    /// keep the default raw-JSON response serializer instead of the HAL serializer.
-    var expectsRawJSONResponse: Bool { get }
-    var body: [AnyHashable: Any] { get }
-    func url(resolver: MSALNativeAuthV2HrefURLResolver) throws -> URL
-}
-
-extension MSALNativeAuthV2Requestable {
-    var httpMethod: String { "POST" }
-    var expectsRawJSONResponse: Bool { false }
-}
-
-/// The destination of a V2 request: either a well-known endpoint or a server-provided HAL `href`.
-enum MSALNativeAuthV2RequestTarget {
-    case endpoint(MSALNativeAuthV2Endpoint)
-    case href(String)
-
-    func url(resolver: MSALNativeAuthV2HrefURLResolver) throws -> URL {
-        switch self {
-        case .endpoint(let endpoint):
-            return try resolver.url(for: endpoint)
-        case .href(let href):
-            return try resolver.url(forHref: href)
-        }
-    }
-}
-
-/// `POST /authorize/challenge` (the authorization challenge that starts a flow). Sends ONLY `client_id` (form encoded).
-struct MSALNativeAuthV2AuthorizeChallengeStartParameters: MSALNativeAuthV2Requestable {
-    let context: MSALNativeAuthRequestContext
-    let clientId: String
-    let encoding: MSALNativeAuthUrlRequestEncoding = .wwwFormUrlEncoded
-    let apiId: MSALNativeAuthTelemetryApiId = .telemetryApiIdV2AuthorizeChallenge
-    let operationType: MSALNativeAuthOperationType = MSALNativeAuthV2OperationType.authorizeChallengeStart.rawValue
-
-    var body: [AnyHashable: Any] {
-        return ["client_id": clientId]
-    }
-
-    func url(resolver: MSALNativeAuthV2HrefURLResolver) throws -> URL {
-        return try resolver.url(for: .authorizeChallenge)
-    }
-}
-
-/// `POST /authorize/challenge` resume. Sends ONLY `continuation_token` (form encoded).
-struct MSALNativeAuthV2AuthorizeChallengeContinueParameters: MSALNativeAuthV2Requestable {
-    let context: MSALNativeAuthRequestContext
-    let continuationToken: String
-    let encoding: MSALNativeAuthUrlRequestEncoding = .wwwFormUrlEncoded
-    let apiId: MSALNativeAuthTelemetryApiId = .telemetryApiIdV2AuthorizeChallenge
-    let operationType: MSALNativeAuthOperationType = MSALNativeAuthV2OperationType.authorizeChallengeContinue.rawValue
-
-    var body: [AnyHashable: Any] {
-        return ["continuation_token": continuationToken]
-    }
-
-    func url(resolver: MSALNativeAuthV2HrefURLResolver) throws -> URL {
-        return try resolver.url(for: .authorizeChallenge)
-    }
-}
-
-/// `POST /token` authorization-code exchange. Form encoded, raw OAuth (non-HAL) response. Includes
-/// `client_info=true` so ESTS returns the `client_info` blob required to persist tokens to the cache.
-struct MSALNativeAuthV2TokenParameters: MSALNativeAuthV2Requestable {
-    let context: MSALNativeAuthRequestContext
-    let clientId: String
-    let code: String
-    let scopes: [String]
-    let encoding: MSALNativeAuthUrlRequestEncoding = .wwwFormUrlEncoded
-    let apiId: MSALNativeAuthTelemetryApiId = .telemetryApiIdV2Token
-    let operationType: MSALNativeAuthOperationType = MSALNativeAuthV2OperationType.token.rawValue
-    let expectsRawJSONResponse = true
-
-    var body: [AnyHashable: Any] {
-        var form: [AnyHashable: Any] = [
-            "grant_type": "authorization_code",
-            "code": code,
-            "client_id": clientId,
-            MSALNativeAuthRequestParametersKey.clientInfo.rawValue: true.description
-        ]
-        if !scopes.isEmpty {
-            form["scope"] = scopes.joined(separator: " ")
-        }
-        return form
-    }
-
-    func url(resolver: MSALNativeAuthV2HrefURLResolver) throws -> URL {
-        return try resolver.url(for: .token)
-    }
-}
-
-/// The signup/signin/resetpassword `start` entry requests. JSON encoded `{username, continuationToken}`,
-/// targeting either the well-known start endpoint or a server-provided `href`.
-struct MSALNativeAuthV2EntryParameters: MSALNativeAuthV2Requestable {
-    let context: MSALNativeAuthRequestContext
-    let target: MSALNativeAuthV2RequestTarget
-    let apiId: MSALNativeAuthTelemetryApiId
-    let operationType: MSALNativeAuthOperationType
-    let username: String
-    let continuationToken: String
-    let encoding: MSALNativeAuthUrlRequestEncoding = .json
-
-    var body: [AnyHashable: Any] {
-        return ["username": username, "continuationToken": continuationToken]
-    }
-
-    func url(resolver: MSALNativeAuthV2HrefURLResolver) throws -> URL {
-        return try target.url(resolver: resolver)
-    }
-}
-
-/// A HAL follow-up request driven by a server-provided `href` (challenge, verify, submit*, register,
-/// update-password, poll). JSON encoded with a caller-supplied body.
-struct MSALNativeAuthV2HrefParameters: MSALNativeAuthV2Requestable {
-    let context: MSALNativeAuthRequestContext
-    let href: String
-    let httpMethod: String
-    let apiId: MSALNativeAuthTelemetryApiId
-    let operationType: MSALNativeAuthOperationType
-    let body: [AnyHashable: Any]
-    let encoding: MSALNativeAuthUrlRequestEncoding = .json
-
-    func url(resolver: MSALNativeAuthV2HrefURLResolver) throws -> URL {
-        return try resolver.url(forHref: href)
-    }
-}
-
-// MARK: - Request configurator
-
 /// Builds a fully-configured `MSIDHttpRequest` for any `MSALNativeAuthV2Requestable`. Subclassing
 /// `MSIDAADRequestConfigurator` gives V2 the standard device-id (`x-client-*`) headers, app metadata,
 /// PkeyAuth, `Accept: application/json`, correlation headers and the authority network-host rewrite.
