+++ MSAL/src/native_auth/network/v2/parameters/MSALNativeAuthV2EntryParameters.swift
@@ -1,0 +1,45 @@
+//
+// Copyright (c) Microsoft Corporation.
+// All rights reserved.
+//
+// This code is licensed under the MIT License.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files(the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and / or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions :
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+import Foundation
+
+/// The signup/signin/resetpassword `start` entry requests. JSON encoded `{username, continuationToken}`,
+/// targeting either the well-known start endpoint or a server-provided `href`.
+struct MSALNativeAuthV2EntryParameters: MSALNativeAuthV2Requestable {
+    let context: MSALNativeAuthRequestContext
+    let target: MSALNativeAuthV2RequestTarget
+    let apiId: MSALNativeAuthTelemetryApiId
+    let operationType: MSALNativeAuthOperationType
+    let username: String
+    let continuationToken: String
+    let encoding: MSALNativeAuthUrlRequestEncoding = .json
+
+    var body: [AnyHashable: Any] {
+        return ["username": username, "continuationToken": continuationToken]
+    }
+
+    func url(resolver: MSALNativeAuthV2HrefURLResolver) throws -> URL {
+        return try target.url(resolver: resolver)
+    }
+}
