+++ azure_pipelines/broker_build_steps.yml
@@ -1,0 +1,149 @@
+# =============================================================================
+# Broker repo-specific build steps (checkouts, submodules, CocoaPods, build).
+#
+# These are the steps that are unique to the Broker submodule validation. The
+# generic macOS tool setup (pool, Xcode, Ruby gems, CocoaPods CLI, xcbeautify)
+# is intentionally NOT here — it is provided by the shared ACES macOS job
+# template (Pipeline YAMLs/shared/aces-macos-job.yml@pipelinesShared), which
+# injects these steps after the tool setup.
+#
+# Parameters:
+#   target   build.py target (e.g. ios_library, mac_library).
+# =============================================================================
+parameters:
+  - name: target
+    type: string
+
+steps:
+  - checkout: azure-activedirectory-tokenbroker-for-objc
+    displayName: 'Checkout Broker'
+    clean: false
+    submodules: false
+    fetchTags: true
+    persistCredentials: true
+
+  - checkout: self
+    displayName: 'Checkout MSAL'
+    clean: false
+    submodules: false
+    fetchTags: true
+    path: 's/azure-activedirectory-tokenbroker-for-objc/ADAuthenticationBroker/Frameworks/microsoft-authentication-library-for-objc'
+    persistCredentials: true
+
+  - task: Bash@3
+    displayName: 'Checkout MSAL submodules + ADAL'
+    inputs:
+      workingDirectory: $(Pipeline.Workspace)/s
+      targetType: 'inline'
+      script: |
+        cd azure-activedirectory-tokenbroker-for-objc
+        git submodule update --init --recursive ADAuthenticationBroker/Frameworks/adal
+        cd ADAuthenticationBroker/Frameworks/microsoft-authentication-library-for-objc
+        git submodule update --init --recursive
+
+  - checkout: WorkplaceJoin-for-iOS
+    displayName: 'Checkout WPJ'
+    clean: false
+    submodules: false
+    fetchTags: true
+    path: 's/azure-activedirectory-tokenbroker-for-objc/ADAuthenticationBroker/Frameworks/WorkplaceJoin-for-iOS'
+    persistCredentials: true
+
+  - task: AzureCLI@2
+    inputs:
+      azureSubscription: 'AuthSdkResourceManager'
+      scriptType: 'pscore'
+      scriptLocation: 'inlineScript'
+      inlineScript: |
+        # if this fails, check out this bash script that includes diagnostics:
+        # https://gist.github.com/johnterickson/19f80a3e969e39f1000d118739176e62
+        # uncomment these for more debugging spew
+        # GIT_TRACE=1
+        # GIT_CURL_VERBOSE=1
+        
+        # Note that the resource is specified to limit the token to Azure DevOps
+        $token = az account get-access-token --query accessToken --resource 499b84ac-1321-427f-aa17-267ca6975798 -o tsv
+        Write-Host "##vso[task.setvariable variable=aadToken;issecret=true]$token"
+
+  - task: Bash@3
+    displayName: 'Checkout NGC Submodules'
+    env:
+      AccessToken: $(MSAzureToken_encoded)
+    inputs:
+      workingDirectory: $(Pipeline.Workspace)/s
+      targetType: 'inline'
+      script: |
+        cd azure-activedirectory-tokenbroker-for-objc/ADAuthenticationBroker/Frameworks
+        git -c http.https://msazure.visualstudio.com/DefaultCollection/One/_git/AD-MFA-NGCAuthentication.extraheader="AUTHORIZATION: bearer $(aadToken)" submodule update --init AD-MFA-NGCAuthentication
+        cd AD-MFA-NGCAuthentication
+        git -c http.https://msazure.visualstudio.com/DefaultCollection/One/_git/AD-MFA-NGCKeyProvider-ios.extraheader="AUTHORIZATION: bearer $(aadToken)" submodule update --init NGCKeyProvider
+        git -c http.https://msazure.visualstudio.com/DefaultCollection/One/_git/AD-MFA-MSAuthNetworking.extraheader="AUTHORIZATION: bearer $(aadToken)" submodule update --init MSAuthNetworking
+
+  - task: Bash@3
+    displayName: 'Checkout WPJ openssl-msft submodule'
+    inputs:
+      workingDirectory: $(Pipeline.Workspace)/s
+      targetType: 'inline'
+      script: |
+        cd azure-activedirectory-tokenbroker-for-objc/ADAuthenticationBroker/Frameworks/WorkplaceJoin-for-iOS
+        git -c http.https://msazure.visualstudio.com/DefaultCollection/PlatformCrypto/_git/openssl-msft.extraheader="AUTHORIZATION: bearer $(aadToken)" submodule update --init Frameworks/openssl-msft
+
+  - task: Bash@3
+    displayName: 'Update WPJ submodules'
+    inputs:
+      workingDirectory: $(Pipeline.Workspace)/s
+      targetType: 'inline'
+      script: |
+        cd azure-activedirectory-tokenbroker-for-objc/ADAuthenticationBroker/Frameworks/WorkplaceJoin-for-iOS
+        git submodule update --init --recursive Frameworks/microsoft-authentication-library-for-objc
+
+  - ${{ if eq(parameters.target, 'mac_library') }}:
+    - task: Cache@2
+      displayName: 'Cache CocoaPods'
+      inputs:
+        key: 'cocoapods | "$(Agent.OS)" | azure-activedirectory-tokenbroker-for-objc/ADAuthenticationBroker/Podfile.lock'
+        path: '$(Pipeline.Workspace)/s/azure-activedirectory-tokenbroker-for-objc/ADAuthenticationBroker/Pods'
+
+    - task: Bash@3
+      displayName: 'Install CocoaPods dependencies'
+      env:
+        AAD_TOKEN: $(aadToken)
+      inputs:
+        workingDirectory: $(Pipeline.Workspace)/s/azure-activedirectory-tokenbroker-for-objc/ADAuthenticationBroker
+        targetType: 'inline'
+        script: |
+          export LANG=en_US.UTF-8
+          export LC_ALL=en_US.UTF-8
+          GIT_CONFIG_COUNT=1 \
+          GIT_CONFIG_KEY_0=http.https://office.visualstudio.com/.extraheader \
+          GIT_CONFIG_VALUE_0="AUTHORIZATION: bearer $AAD_TOKEN" \
+          pod install
+      retryCountOnTaskFailure: 1
+
+  - task: UsePythonVersion@0
+    displayName: 'Use Python 3.x'
+
+  - task: Bash@3
+    displayName: 'Run a python script for Broker'
+    inputs:
+      targetType: 'inline'
+      script: |
+        cd azure-activedirectory-tokenbroker-for-objc
+        echo "executing build:./build.py"
+        { output=$(./build.py --show-build-settings --target ${{ parameters.target }} 2>&1 1>&3-) ;} 3>&1
+        final_status=$(<./build/status.txt)
+        echo "FINAL STATUS  = ${final_status}"
+        echo "POSSIBLE ERRORS: ${output}"
+
+        if [ $final_status != "0" ]; then
+          echo "Build & Testing Failed! \n ${output}" >&2
+        fi
+      failOnStderr: true
+
+  - task: Bash@3
+    condition: always()
+    displayName: Cleanup
+    inputs:
+      targetType: 'inline'
+      script: |
+        rm -rf ./build/status.txt
